+++ contracts/SecurityTokenRegistry.sol
@@ -60,7 +60,7 @@ contract SecurityTokenRegistry is ISecurityTokenRegistry, Util, Pausable, Regist
             polymathRegistry
         );
 
-        securityTokens[newSecurityTokenAddress] = SecurityTokenData(symbol, _tokenDetails);
+        securityTokens[newSecurityTokenAddress] = SecurityTokenData(symbol, _tokenDetails, now);
         symbols[symbol] = newSecurityTokenAddress;
         emit LogNewSecurityToken(symbol, newSecurityTokenAddress, msg.sender);
     }
@@ -81,7 +81,7 @@ contract SecurityTokenRegistry is ISecurityTokenRegistry, Util, Pausable, Regist
         require(_owner != address(0));
         require(!(ITickerRegistry(tickerRegistry).isReserved(symbol, _owner, _name, _swarmHash)), "Trying to use non-valid symbol");
         symbols[symbol] = _securityToken;
-        securityTokens[_securityToken] = SecurityTokenData(symbol, _tokenDetails);
+        securityTokens[_securityToken] = SecurityTokenData(symbol, _tokenDetails, now);
         emit LogAddCustomSecurityToken(_name, symbol, _securityToken, now);
     }
 
@@ -110,16 +110,18 @@ contract SecurityTokenRegistry is ISecurityTokenRegistry, Util, Pausable, Regist
 
      /**
      * @notice Get security token data by its address
-     * @param _securityToken Address of the Scurity token
-     * @return string
-     * @return address
-     * @return string
+     * @param _securityToken Address of the Scurity token.
+     * @return string Symbol of the Security Token.
+     * @return address Address of the issuer of Security Token.
+     * @return string Details of the Token.
+     * @return uint256 Timestamp at which Security Token get launched on Polymath platform.
      */
-    function getSecurityTokenData(address _securityToken) public view returns (string, address, string) {
+    function getSecurityTokenData(address _securityToken) public view returns (string, address, string, uint256) {
         return (
             securityTokens[_securityToken].symbol,
             ISecurityToken(_securityToken).owner(),
-            securityTokens[_securityToken].tokenDetails
+            securityTokens[_securityToken].tokenDetails,
+            securityTokens[_securityToken].registrationTimestamp
         );
     }